+++ applications/GeoMechanicsApplication/custom_elements/Pw_element.h
@@ -18,7 +18,7 @@
 #include "calculation_contribution.h"
 #include "compressibility_calculator.hpp"
 #include "custom_retention/retention_law_factory.h"
-#include "custom_utilities/check_utilities.h"
+#include "custom_utilities/check_utilities.hpp"
 #include "custom_utilities/constitutive_law_utilities.h"
 #include "custom_utilities/dof_utilities.h"
 #include "custom_utilities/element_utilities.hpp"
