+++ applications/GeoMechanicsApplication/custom_elements/geo_steady_state_Pw_piping_element.h
@@ -14,7 +14,7 @@
 
 #include <numeric>
 
-#include "custom_utilities/check_utilities.h"
+#include "custom_utilities/check_utilities.hpp"
 #include "custom_utilities/dof_utilities.h"
 #include "custom_utilities/element_utilities.hpp"
 #include "custom_utilities/math_utilities.h"
