+++ applications/GeoMechanicsApplication/custom_elements/transient_thermal_element.h
@@ -17,7 +17,7 @@
 #include "custom_constitutive/thermal_dispersion_law.h"
 #include "custom_constitutive/thermal_filter_law.h"
 #include "custom_retention/retention_law_factory.h"
-#include "custom_utilities/check_utilities.h"
+#include "custom_utilities/check_utilities.hpp"
 #include "custom_utilities/dof_utilities.h"
 #include "custom_utilities/variables_utilities.hpp"
 #include "geo_mechanics_application_variables.h"
