+++ applications/GeoMechanicsApplication/custom_utilities/check_utilities.cpp
@@ -12,7 +12,7 @@
 //
 
 // Project includes
-#include "check_utilities.h"
+#include "check_utilities.hpp"
 #include "geo_mechanics_application_variables.h"
 #include "includes/exception.h"
 
@@ -20,6 +20,34 @@
 
 namespace Kratos
 {
+
+CheckProperties::CheckProperties(const Properties&       rProperties,
+                                 const std::string&      rPrintName,
+                                 CheckProperties::Bounds RangeBoundsType)
+    : mrProperties(rProperties), mrPrintName(rPrintName), mRangeBoundsType(RangeBoundsType)
+{
+}
+
+CheckProperties::CheckProperties(const Properties&       rProperties,
+                                 const std::string&      rPrintName,
+                                 std::size_t             ElementId,
+                                 CheckProperties::Bounds RangeBoundsType)
+    : mrProperties(rProperties), mrPrintName(rPrintName), mElementId(ElementId), mRangeBoundsType(RangeBoundsType)
+{
+}
+
+CheckProperties CheckProperties::SingleUseBounds(CheckProperties::Bounds RangeBoundsType) const
+{
+    if (mElementId) {
+        return CheckProperties(mrProperties, mrPrintName, *mElementId, RangeBoundsType);
+    }
+    return CheckProperties(mrProperties, mrPrintName, RangeBoundsType);
+}
+
+void CheckProperties::SetNewRangeBounds(CheckProperties::Bounds RangeBoundsType) const
+{
+    mRangeBoundsType = RangeBoundsType;
+}
 
 void CheckUtilities::CheckDomainSize(double DomainSize, std::size_t Id, const std::optional<std::string>& PrintName)
 {
@@ -100,5 +128,26 @@ void CheckUtilities::CheckForNonZeroZCoordinateIn2D(const Geometry<Node>& rGeome
     });
     KRATOS_ERROR_IF_NOT(pos == rGeometry.end())
         << "Node with Id: " << pos->Id() << " has non-zero Z coordinate." << std::endl;
+}
+
+std::string CheckProperties::double_to_string(double value) const
+{
+    std::ostringstream oss;
+    oss << std::defaultfloat << value;
+    return oss.str();
+}
+
+std::string CheckProperties::print_property_id() const
+{
+    std::ostringstream oss;
+    oss << " in the " << mrPrintName << " with Id " << mrProperties.Id();
+    return oss.str();
+}
+
+std::string CheckProperties::print_element_id() const
+{
+    std::ostringstream oss;
+    if (mElementId) oss << " at element with Id " << *mElementId;
+    return oss.str();
 }
 } /* namespace Kratos.*/
